+++ src/test/ts/browser/LoadTinyTest.ts
@@ -12,8 +12,8 @@ UnitTest.asynctest('LoadTinyTest', (success, failure) => {
 
     delete Global.tinymce;
     delete Global.tinyMCE;
-
-    const hasTinymceUri = (attrName: string) => (elm: SugarElement) => Attribute.getOpt(elm, attrName).exists((src) => Strings.contains(src, 'tinymce'));
+    const hasTinymceUri = (attrName: string) => (elm: SugarElement<Element>) =>
+      Attribute.getOpt(elm, attrName).exists((src) => Strings.contains(src, 'tinymce'));
 
     const elements = Arr.flatten([
       Arr.filter(SelectorFilter.all('script'), hasTinymceUri('src')),